requirements.txt
streamlit
numpy
opencv-python
opencv-python-headless
tensorflow
keras
collection
google-generativeai